pyproject.toml
[project]
name = "gov-program-predictor"
version = "0.1.0"
description = "Government Program Predictor using LangChain"
authors = [
    {name = "Chris Fabian"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "langchain (>=0.3.15,<0.4.0)",
    "langchain-openai (>=0.3.1,<0.4.0)",
    "langchain-community (>=0.3.15,<0.4.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
